backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyPDF2==3.0.1
chromadb
numpy
torch
transformers
groq
sqlalchemy==2.0.23
alembic==1.13.1